content/canvas/test/test_2d.canvas.readonly.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.canvas.readonly</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
var _deferred = false;

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var c = document.createElement('canvas');
var d = ctx.canvas;
try { ctx.canvas = c; } catch (e) {} // not specified whether this throws, I think
ok(ctx.canvas === d, "ctx.canvas === d");

if (!_deferred) SimpleTest.finish();
});
</script>
